requirements.txt
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
pika==1.2.0
psycopg2==2.9.1
